:recycle: refactor(api/endpoint): Extract authorized/Authentication.
