+++ imports/api/endpoint/Params.ts
@@ -1,4 +1,4 @@
-import Authentication from './Authentication';
+import Authentication from '../Authentication';
 import Executor from './Executor';
 import Options from './Options';
 import Transaction from './Transaction';
